Reorganized events files
