+++ TaaontiaCore/Events/FightEvent.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TaaontiaCore.Events.Fight
+namespace TaaontiaCore.Events
 {
     public class FightEvent
     {
